% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{biased_trans_matrix}
\alias{biased_trans_matrix}
\title{Generate Biased Transition Matrix}
\usage{
biased_trans_matrix(data, all_data = FALSE)
}
\arguments{
\item{data}{a data.frame containing mosly categorical data}

\item{all_data}{a boolean (default: FALSE) on whether to return additional node and edge data (see Note)}
}
\value{
a biased transition matrix of dim [k,k] where k is the number of unique feature values
}
\section{Note}{

If all_data is `TRUE`, A list with three members will be returned:
\itemize{
   \item \strong{nodes} a tibble containing all relevant information for each node, or feature value
   \item \strong{edges} a tibble containing all relevant information for each edge, or feature value coupling
   \item \strong{Wb} the biased transition matrix
}
}

