% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{categorical_to_int}
\alias{categorical_to_int}
\title{Helper fuction to transform a categorical dataset
into one represented by integers}
\usage{
categorical_to_int(data)
}
\arguments{
\item{data}{a tibble of categorical data}
}
\value{
a list containing:
\itemize{
   \item character vector of all unique feature values (V)
   \item converted tibble with character values represented as integers
}
}
\description{
Helper fuction to transform a categorical dataset
into one represented by integers
}
