% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{counting_helpers}
\alias{counting_helpers}
\alias{intra_freq}
\alias{inter_freq}
\title{Counting helper functions}
\usage{
intra_freq(data)

inter_freq(data)
}
\arguments{
\item{data}{a tibble of categorical data}

\item{data}{a tibble of categorical data}
}
\value{
a tibble containing all unique combinations of feature values
   with columns: \emph{u}, \emph{v}, \emph{freq}, emph{p}, and \emph{feature}

a tibble containing all unique combinations of feature values
   with columns: \emph{u}, \emph{v}, \emph{freq}, and \emph{group}
}
\description{
Helper function to calculate all bivariate frequencies
within the dataset
}
