% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{feature_relevance}
\alias{feature_relevance}
\title{Display the computed feature relevance in the CBRW model}
\usage{
feature_relevance(data)
}
\arguments{
\item{data}{A tibble as provided by the output of a cbrw call}
}
\value{
a tibble containing each feature with it's computed relevance
}
\description{
TODO
}
